+++ tests/unit/test_api.py
@@ -4,7 +4,14 @@ from uuid import UUID
 
 import pytest
 
-from ctms.crud import get_contacts_by_any_id
+from ctms.crud import (
+    create_contact,
+    get_amo_by_email_id,
+    get_contacts_by_any_id,
+    get_fxa_by_email_id,
+    get_newsletters_by_email_id,
+    get_vpn_by_email_id,
+)
 from ctms.models import Email
 from ctms.sample_data import SAMPLE_CONTACTS
 from ctms.schemas import ContactInSchema, ContactSchema
@@ -333,6 +340,65 @@ def test_get_ctms_by_alt_id_none_found(client, dbsession, alt_id_name, alt_id_va
     assert resp.status_code == 200
     data = resp.json()
     assert len(data) == 0
+
+
+@pytest.fixture
+def add_contact(request, client, dbsession):
+    _id = (
+        request.param
+        if hasattr(request, "param")
+        else "d1da1c99-fe09-44db-9c68-78a75752574d"
+    )
+    email_id = UUID(str(_id))
+    contact = SAMPLE_CONTACTS[email_id]
+    fields_not_written = SAMPLE_CONTACTS.get_not_written(email_id)
+
+    def _add(
+        modifier: Callable[[ContactSchema], ContactSchema] = lambda x: x,
+        code: int = 303,
+        stored_contacts: int = 1,
+        check_redirect: bool = True,
+        query_fields: dict = {"primary_email": contact.email.primary_email},
+        check_written: bool = True,
+    ):
+        sample = contact.copy(deep=True)
+        sample = modifier(sample)
+        resp = client.post("/ctms", sample.json())
+        assert resp.status_code == code, resp.text
+        if check_redirect:
+            assert resp.headers["location"] == f"/ctms/{sample.email.email_id}"
+        saved = [
+            ContactSchema(**c)
+            for c in get_contacts_by_any_id(dbsession, **query_fields)
+        ]
+        assert len(saved) == stored_contacts
+
+        # Now make sure that we skip writing default models
+        def _check_written(field, getter):
+            results = getter(dbsession, sample.email.email_id)
+            if sample.dict()[field] and code == 303:
+                if field in fields_not_written:
+                    assert (
+                        results is None
+                    ), f"{email_id} has field `{field}` but it is _default_ and it should _not_ have been written to db"
+                else:
+                    assert (
+                        results
+                    ), f"{email_id} has field `{field}` and it should have been written to db"
+            else:
+                assert (
+                    results is None
+                ), f"{email_id} does not have field `{field}` and it should _not_ have been written to db"
+
+        if check_written:
+            _check_written("amo", get_amo_by_email_id)
+            _check_written("fxa", get_fxa_by_email_id)
+            _check_written("newsletters", get_newsletters_by_email_id)
+            _check_written("vpn_waitlist", get_vpn_by_email_id)
+
+        return saved, sample, email_id
+
+    return _add
 
 
 def _compare_written_contacts(
